+++ src/main/kotlin/cloud/glitchdev/rfu/config/RFUSettings.kt
@@ -4,6 +4,7 @@ import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.config.categories.BackendSettings
 import cloud.glitchdev.rfu.config.categories.DevSettings
 import cloud.glitchdev.rfu.config.categories.GeneralFishing
+import cloud.glitchdev.rfu.config.categories.HotSpotSettings
 import cloud.glitchdev.rfu.config.categories.LavaFishing
 import cloud.glitchdev.rfu.config.categories.OtherSettings
 import cloud.glitchdev.rfu.config.categories.RareScSettings
@@ -126,6 +127,7 @@ object RFUSettings : ConfigKt("rfu/settings") {
         category(GeneralFishing)
         category(RareScSettings)
         category(LavaFishing)
+        category(HotSpotSettings)
         category(OtherSettings)
         category(BackendSettings)
         category(DevSettings)