+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Changes:
 - Merged XP/h and SC/h displays into a single Fish Tracking display.
-- Unified fishing tracking settings.
+- Unified fishing tracking settings and added a migration step.
 - Made the "overall" part of the fishing tracker more compact.
+- Implemented Highlight Hotspot feature with radius detection and buff-based coloring.
 
 ### Fixes:
 - Fixed description of failed cast sound